# frontend/src/api/openapi.json
{
  "components": {
    "schemas": {
...
      "ContextDocIn": {
        "additionalProperties": false,
        "description": "A document inlined into the prompts. Bytes only — no path ever reaches the CLI.",
        "properties": {
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "text": {
            "maxLength": 200000,
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "name",
          "text"
        ],
        "title": "ContextDocIn",
        "type": "object"
      },
      "ContextDocRef": {
        "properties": {
          "cap_chars": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cap Chars"
          },
          "chars": {
            "title": "Chars",
            "type": "integer"
          },
          "delivered": {
            "default": "pending",
            "enum": [
              "full",
              "truncated",
              "omitted",
              "pending"
            ],
            "title": "Delivered",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "cap_chars",
          "chars",
          "delivered",
          "name"
        ],
        "title": "ContextDocRef",
        "type": "object"
      },
...
      "CreateRunRequest": {
        "additionalProperties": false,
        "properties": {
          "config": {
            "$ref": "#/components/schemas/RunConfigIn"
          },
          "context_docs": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/ContextDocIn"
            },
            "maxItems": 5,
            "title": "Context Docs",
            "type": "array"
          },
          "from_run": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "From Run"
          },
          "harness": {
            "default": "claude",
            "enum": [
              "claude",
              "codex",
              "demo"
            ],
            "title": "Harness",
            "type": "string"
          },
          "prompt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt"
          },
          "question": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Question"
          },
          "title": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "title": "CreateRunRequest",
        "type": "object"
      },
...
      "EventTicket": {
        "properties": {
          "expires_in": {
            "title": "Expires In",
            "type": "integer"
          },
          "ticket": {
            "title": "Ticket",
            "type": "string"
          }
        },
        "required": [
          "expires_in",
          "ticket"
        ],
        "title": "EventTicket",
        "type": "object"
      },
      "FeedbackEntry": {
        "properties": {
          "guidance": {
            "title": "Guidance",
            "type": "string"
          },
          "round": {
            "title": "Round",
            "type": "integer"
          }
        },
        "required": [
          "guidance",
          "round"
        ],
        "title": "FeedbackEntry",
        "type": "object"
      },
...
      "GraftEvent": {
        "properties": {
          "abstained_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Abstained Reason"
          },
          "fired": {
            "title": "Fired",
            "type": "boolean"
          },
          "hhi": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hhi"
          },
          "n_clusters": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "N Clusters"
          },
          "round": {
            "title": "Round",
            "type": "integer"
          },
          "seed_framing": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seed Framing"
          },
          "source_domain": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Domain"
          },
          "votes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Votes"
          }
        },
        "required": [
          "abstained_reason",
          "fired",
          "hhi",
          "n_clusters",
          "round",
          "seed_framing",
          "source_domain",
          "votes"
        ],
        "title": "GraftEvent",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "required": [
          "detail"
        ],
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Harness": {
        "enum": [
          "claude",
          "codex",
          "demo"
        ],
        "title": "Harness",
        "type": "string"
      },
...
      "ModelTableEntry": {
        "properties": {
          "effort": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effort"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "role": {
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "effort",
          "model",
          "role"
        ],
        "title": "ModelTableEntry",
        "type": "object"
      },
...
      "RunBudget": {
        "properties": {
          "budget_calls": {
            "title": "Budget Calls",
            "type": "integer"
          },
          "budget_usd": {
            "title": "Budget Usd",
            "type": "number"
          },
          "by_role": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/BudgetByRole"
            },
            "title": "By Role",
            "type": "array"
          },
          "calls_used": {
            "title": "Calls Used",
            "type": "integer"
          },
          "spend_usd": {
            "title": "Spend Usd",
            "type": "number"
          }
        },
        "required": [
          "budget_calls",
          "budget_usd",
          "by_role",
          "calls_used",
          "spend_usd"
        ],
        "title": "RunBudget",
        "type": "object"
      },
      "RunConfigIn": {
        "description": "Plan C5's config block. `model_table` is not accepted: the launcher resolves it.",
        "properties": {
          "budget_calls": {
            "default": 150,
            "exclusiveMinimum": 0.0,
            "maximum": 10000.0,
            "title": "Budget Calls",
            "type": "integer"
          },
          "budget_usd": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "maximum": 1000.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Budget Usd"
          },
          "evolve_top_k": {
            "default": 3,
            "maximum": 25.0,
            "minimum": 0.0,
            "title": "Evolve Top K",
            "type": "integer"
          },
          "generation_batch": {
            "default": 8,
            "maximum": 40.0,
            "minimum": 1.0,
            "title": "Generation Batch",
            "type": "integer"
          },
          "graft": {
            "$ref": "#/components/schemas/GraftConfigIn"
          },
          "grounding_depth": {
            "$ref": "#/components/schemas/GroundingDepth",
            "default": "standard"
          },
          "matches_per_round": {
            "default": 6,
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Matches Per Round",
            "type": "integer"
          },
          "model_overrides": {
            "anyOf": [
              {
                "additionalProperties": {
                  "$ref": "#/components/schemas/ModelOverrideIn"
                },
                "propertyNames": {
                  "enum": [
                    "generation",
                    "reflection",
                    "proximity",
                    "ranking",
                    "evolution",
                    "meta_review",
                    "cartographer",
                    "overview",
                    "workshop",
                    "framing",
                    "verification",
                    "synthesis",
                    "challenge"
                  ]
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Overrides"
          },
          "model_tier": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ModelTier"
              },
              {
                "type": "null"
              }
            ]
          },
          "provider": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Provider"
              },
              {
                "type": "null"
              }
            ]
          },
          "rounds": {
            "default": 5,
            "maximum": 25.0,
            "minimum": 1.0,
            "title": "Rounds",
            "type": "integer"
          },
          "runner": {
            "$ref": "#/components/schemas/RunnerKind",
            "default": "claude"
          },
          "wall_clock_minutes": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "maximum": 1440.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wall Clock Minutes"
          },
          "workflow": {
            "default": "adaptive",
            "enum": [
              "adaptive",
              "tournament"
            ],
            "title": "Workflow",
            "type": "string"
          }
        },
        "title": "RunConfigIn",
        "type": "object"
      },
...
      "RunDetail": {
        "properties": {
          "budget": {
            "$ref": "#/components/schemas/RunBudget"
          },
          "config": {
            "additionalProperties": true,
            "default": {},
            "title": "Config",
            "type": "object"
          },
          "context_docs": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/ContextDocRef"
            },
            "title": "Context Docs",
            "type": "array"
          },
          "degraded_count": {
            "default": 0,
            "title": "Degraded Count",
            "type": "integer"
          },
          "failed_calls": {
            "default": 0,
            "title": "Failed Calls",
            "type": "integer"
          },
          "feedback_history": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/FeedbackEntry"
            },
            "title": "Feedback History",
            "type": "array"
          },
          "graft_events": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/GraftEvent"
            },
            "title": "Graft Events",
            "type": "array"
          },
          "leaderboard": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/HypothesisRow"
            },
            "title": "Leaderboard",
            "type": "array"
          },
          "lost_steps": {
            "default": 0,
            "title": "Lost Steps",
            "type": "integer"
          },
          "model_table": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/ModelTableEntry"
            },
            "title": "Model Table",
            "type": "array"
          },
          "problems": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/Event"
            },
            "title": "Problems",
            "type": "array"
          },
          "recent_events": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/Event"
            },
            "title": "Recent Events",
            "type": "array"
          },
          "research": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResearchView"
              },
              {
                "type": "null"
              }
            ]
          },
          "retried_calls": {
            "default": 0,
            "title": "Retried Calls",
            "type": "integer"
          },
          "rounds": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/RoundSummary"
            },
            "title": "Rounds",
            "type": "array"
          },
          "run": {
            "$ref": "#/components/schemas/RunSummary"
          }
        },
        "required": [
          "budget",
          "config",
          "context_docs",
          "degraded_count",
          "failed_calls",
          "feedback_history",
          "graft_events",
          "leaderboard",
          "lost_steps",
          "model_table",
          "problems",
          "recent_events",
          "research",
          "retried_calls",
          "rounds",
          "run"
        ],
        "title": "RunDetail",
        "type": "object"
      },
...
      "RunList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/RunSummary"
            },
            "title": "Items",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total"
        ],
        "title": "RunList",
        "type": "object"
      },
      "RunSource": {
        "enum": [
          "app",
          "imported"
        ],
        "title": "RunSource",
        "type": "string"
      },
      "RunSummary": {
        "properties": {
          "archived": {
            "title": "Archived",
            "type": "boolean"
          },
          "budget_calls": {
            "title": "Budget Calls",
            "type": "integer"
          },
          "calls_used": {
            "title": "Calls Used",
            "type": "integer"
          },
          "counts": {
            "$ref": "#/components/schemas/RunCounts"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "elapsed_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Elapsed Seconds"
          },
          "ended_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended Reason"
          },
          "engine_run_id": {
            "title": "Engine Run Id",
            "type": "string"
          },
          "failed_calls": {
            "default": 0,
            "title": "Failed Calls",
            "type": "integer"
          },
          "graft": {
            "$ref": "#/components/schemas/RunGraft"
          },
          "harness": {
            "$ref": "#/components/schemas/Harness"
          },
          "has_overview": {
            "title": "Has Overview",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "lifecycle": {
            "$ref": "#/components/schemas/Lifecycle"
          },
          "lost_steps": {
            "default": 0,
            "title": "Lost Steps",
            "type": "integer"
          },
          "model_level": {
            "anyOf": [
              {
                "enum": [
                  "low",
                  "med",
                  "high",
                  "max"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Level"
          },
          "model_level_custom": {
            "default": false,
            "title": "Model Level Custom",
            "type": "boolean"
          },
          "overview_skipped_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Overview Skipped Reason"
          },
          "owner_display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Display Name"
          },
          "question": {
            "title": "Question",
            "type": "string"
          },
          "round": {
            "title": "Round",
            "type": "integer"
          },
          "rounds_target": {
            "title": "Rounds Target",
            "type": "integer"
          },
          "source": {
            "$ref": "#/components/schemas/RunSource"
          },
          "source_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Version"
          },
          "spend_usd": {
            "title": "Spend Usd",
            "type": "number"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "tokens_total": {
            "title": "Tokens Total",
            "type": "integer"
          },
          "top": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/RunTop"
            },
            "title": "Top",
            "type": "array"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "archived",
          "budget_calls",
          "calls_used",
          "counts",
          "created_at",
          "elapsed_seconds",
          "ended_reason",
          "engine_run_id",
          "failed_calls",
          "graft",
          "harness",
          "has_overview",
          "id",
          "lifecycle",
          "lost_steps",
          "model_level",
          "model_level_custom",
          "overview_skipped_reason",
          "owner_display_name",
          "question",
          "round",
          "rounds_target",
          "source",
          "source_version",
          "spend_usd",
          "title",
          "tokens_total",
          "top",
          "updated_at"
        ],
        "title": "RunSummary",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "ctx",
          "input",
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Local-first API for Oracle run orchestration and artifact discovery.",
    "title": "Oracle GUI API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/runs": {
      "get": {
        "description": "A page of runs, newest first.\n\nDemo runs are hidden unless asked for, by `include_demo` or by naming the harness:\nthey are practice runs and would otherwise crowd out the scientist's real work.",
        "operationId": "list_runs_api_runs_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "harness",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Harness"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "show_archived",
            "required": false,
            "schema": {
              "default": false,
              "title": "Show Archived",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "include_demo",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Demo",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "default": "recent",
              "enum": [
                "recent",
                "oldest",
                "title",
                "calls"
              ],
              "title": "Sort",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 100,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "mine",
            "required": false,
            "schema": {
              "default": false,
              "title": "Mine",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Runs",
        "tags": [
          "runs"
        ]
      },
      "post": {
        "description": "Create a run and start its supervisor. Returns as soon as the process is alive.\n\nAlways a **new** run, `from_run` included: that field copies an earlier run's question,\nprompt, config and context documents, and the run it creates starts from no hypotheses,\nno ratings and no guidance. Extending a run that already exists is\n`POST /runs/{id}/controls {action: \"continue\"}`, which keeps the whole idea pool.\n\nThe 409 is the harness lane: one active run per harness, enforced by a partial unique\nindex rather than a check the API could race against. Its details name the run holding\nthe lane only when that run is visible to this user; otherwise they report capacity\nwithout exposing another workspace.",
        "operationId": "create_run_api_runs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Run",
        "tags": [
          "runs"
        ]
      }
    },
...
    "/api/runs/{run_id}/events": {
      "get": {
        "description": "Every event of this run from `after_seq` on, as it happens.",
        "operationId": "stream_run_events_api_runs__run_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "after_seq",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "After Seq",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stream Run Events",
        "tags": [
          "events"
        ]
      }
    },
    "/api/runs/{run_id}/events/ticket": {
      "post": {
        "description": "A credential good for one stream of this run, for the next minute.\n\nThe ticket is bound to the id exactly as spelled in this URL, since the middleware\nthat redeems it compares path segments and never resolves a run.",
        "operationId": "issue_event_ticket_api_runs__run_id__events_ticket_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventTicket"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Issue Event Ticket",
        "tags": [
          "events"
        ]
      }
    },
...
  }
}